move channel typedef into portability layer
